+++ CHANGES.md
@@ -53,6 +53,7 @@
 * Use timezones to check unaired status during show update/adding
 * Fix syntax error causing renamer to error out
 * Change storing metadata nfo vars from int to strings to resolve lxml type exceptions that don't occur with etree
+* Add visual indicator for upcoming or started shows on Add Browse Shows
 
 
 ### 0.11.11 (2016-04-05 19:20:00 UTC)